refactor: update field component structure
