+++ docs/components/FieldImageEdit.md
@@ -1,5 +1,5 @@
-`FieldImage`
-============
+`FieldImageEdit`
+================
