+++ docs/components/FieldNumberEdit.md
@@ -1,7 +1,7 @@
-`FieldString`
-=============
+`FieldNumberEdit`
+=================
 
-Component for API fields of type String
+Component for API fields of type Number.
 
 Props
 -----
@@ -46,7 +46,6 @@ The ID of the document being edited.
 If defined, contains an error message to be displayed by the field.
 
 - type: `string`
-- default value: `false`
 
 
 ### `forceValidation`
@@ -55,7 +54,6 @@ Whether the field should be validated as soon as it mounts, rather than
 waiting for a change event.
 
 - type: `bool`
-- default value: `false`
 
 
 ### `group`
@@ -96,5 +94,4 @@ The field schema.
 The field value.
 
 - type: `bool`
-- default value: `''`
