+++ docs/components/FileUploadEdit.md
@@ -1,5 +1,5 @@
-`FileUpload`
-============
+`FileUploadEdit`
+================
