+++ frontend/components/FileUpload/FieldUpload.jsx
@@ -1,0 +1,3 @@
+import edit from './FieldUploadEdit'
+
+export {edit}
