+++ webpack.config.js
@@ -126,7 +126,7 @@ module.exports = {
       let componentsFile = '// Generated by Webpack. Do not change or commit!\n'
 
       componentsFile += components.map(component => {
-        return `import ${component} from '${PATHS.COMPONENTS}/${component}/${component}'`
+        return `import * as ${component} from '${PATHS.COMPONENTS}/${component}/${component}'`
       }).join('\n')
 
       componentsFile += `\n\nexport {${components.join(', ')}}`
